+++ CHANGELOG.md
@@ -16,7 +16,10 @@
   - `array_except`
   - `create_map`
   - `sign`/`signum`
-- Added moving_agg function in DataFrame.analytics for enabling moving aggregations like sums and averages with multiple window sizes.
+- Added following functions to DataFrame.analytics
+  - Added moving_agg function in DataFrame.analytics for enabling moving aggregations like sums and averages with multiple window sizes.
+  - Added cummulative_agg function in DataFrame.analytics for enabling commulative aggregations like sums and averages on multiple columns.
+  - Added compute_lag and compute_lead function in DataFrame.analytics for enabling lead and lag calculations on multiple columns.
 
 ### Bug Fixes